skills/rag-pipeline/evals/iteration-1/eval-1-without_skill/outputs/requirements.txt
pyzotero==1.5.22
langchain==0.1.20
langchain-community==0.0.36
langchain-openai==0.1.15
langchain-text-splitters==0.0.2
pypdf==4.0.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pgvector==0.2.4
numpy==1.24.3
scikit-learn==1.3.2
requests==2.31.0
pydantic==2.5.3
tqdm==4.66.2
